병합

+++ src/main/java/com/apartment/apart/domain/community/CommunityController.java
@@ -32,7 +32,7 @@ public class CommunityController {
     }
     @PreAuthorize("isAuthenticated()")
     @GetMapping("/detail/{id}")
-    public String detail(Model model, @PathVariable("id") Integer id, CommunityReplyForm communityReplyForm) {
+    public String detail(Model model, @PathVariable("id") Long id, CommunityReplyForm communityReplyForm) {
         Community community = this.communityService.getCommunity(id);
         model.addAttribute("community", community);
         return "community_detail";
@@ -55,7 +55,7 @@ public class CommunityController {
 
     @PreAuthorize("isAuthenticated()")
     @GetMapping("/modify/{id}")
-    public String communityModify(CommunityForm communityForm, @PathVariable("id") Integer id, Principal principal) {
+    public String communityModify(CommunityForm communityForm, @PathVariable("id") Long id, Principal principal) {
         Community community = this.communityService.getCommunity(id);
         if(!community.getUser().getUserId().equals(principal.getName())) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "수정권한이 없습니다.");
@@ -68,7 +68,7 @@ public class CommunityController {
     @PreAuthorize("isAuthenticated()")
     @PostMapping("/modify/{id}")
     public String communityModify(@Valid CommunityForm communityForm, BindingResult bindingResult,
-                               Principal principal, @PathVariable("id") Integer id) {
+                                  Principal principal, @PathVariable("id") Long id) {
         if (bindingResult.hasErrors()) {
             return "community_form";
         }
@@ -82,7 +82,7 @@ public class CommunityController {
 
     @PreAuthorize("isAuthenticated()")
     @GetMapping("/delete/{id}")
-    public String communityDelete(Principal principal, @PathVariable("id") Integer id) {
+    public String communityDelete(Principal principal, @PathVariable("id") Long id) {
         Community community = this.communityService.getCommunity(id);
         if (!community.getUser().getUserId().equals(principal.getName())) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "삭제권한이 없습니다.");
